Restartable sequences for tcmalloc.
General idea:  Use atomic sequences for the fast path of push/pop the linked list.
For the slow path, take a per-cpu lock.

On the slow path the kernel can sleep, so put all the threads requesting memory to sleep also.

The list size is tagged on the upper 16 bits of the linked list head, this could/should be a cmpxchg16b instead?
Every x deallocations, go through and sum up the total heap size to see if we should call Scavenge(), since we can't
atomically set the heap_size_ variable also.

There are also some hacks to try and increase the transfer cache size, I should figure out a better way to do that.